move plotting to pl submodule and updated API docs
